Adding Authentication and Testing

+++ src/main/java/com/practice/journalApp/service/JournalEntryService.java
@@ -33,7 +33,7 @@ public class JournalEntryService {
 		journalEntry.setDate(LocalDateTime.now());
 		JournalEntry saved=journalEntryRepository.save(journalEntry);
 		user.getJournalEntries().add(saved);
-		userService.saveEntry(user);
+		userService.saveUser(user);
 
 	}
 	
@@ -52,12 +52,23 @@ public class JournalEntryService {
 		return journalEntryRepository.findById(id);
 	}
 	
-	public void deleteById(ObjectId id, String userName)
+	@Transactional
+	public boolean deleteById(ObjectId id, String userName)
 	{
 		User user=userService.findByUserName(userName);
-		user.getJournalEntries().removeIf(x->x.getId().equals(id));
-		userService.saveEntry(user);
-		journalEntryRepository.deleteById(id);
+		boolean removed=user.getJournalEntries().removeIf(x->x.getId().equals(id));
+		if(removed)
+		{
+			userService.saveUser(user);
+			journalEntryRepository.deleteById(id);
+			return true;
+		}
+		return false;
+
 	}
+//	public List<JournalEntry> findByUserName(String userName)
+//	{
+//		
+//	}
 	
 }
